Revert "update motor speed stuff"
This reverts commit 1a59460a9f7b6f29c80c4c1cbd59e0b44df03470.

+++ firmware/sensor_microcontroller/src/microRosFunctions.cpp
@@ -182,8 +182,8 @@ bool create_entities() {
   sensorMsg.left_speed = 0;
   sensorMsg.right_speed = 0;
 #elif ROS_DEBUG
-  msg.left_speed = 0.0f;
-  msg.right_speed = 0.0f;
+  msg.left_speed = 0;
+  msg.right_speed = 0;
 #endif
 
   state = WAITING_AGENT;
@@ -325,8 +325,8 @@ boolean microRosSetup(unsigned int timer_timeout, const char *nodeName,
   sensorMsg.left_speed = 0;
   sensorMsg.right_speed = 0;
 #elif ROS_DEBUG
-  msg.left_speed = 0.0f;
-  msg.right_speed = 0.0f;
+  msg.left_speed = 0;
+  msg.right_speed = 0;
 #endif
   return true;
 }
@@ -379,9 +379,8 @@ void transmitMsg(RefDisplacement thetaRef, RefSpeed omegaRef,
                  FanSpeeds fanSpeeds, IMUData imuData) {
   sensorMsg.long_disp = thetaRef.longDisp;
   sensorMsg.lat_disp = thetaRef.latDisp;
-  // Sensors.msg keeps left/right_speed as int8 — narrow the float command here.
-  sensorMsg.left_speed = static_cast<int8_t>(omegaRef.leftSpeed);
-  sensorMsg.right_speed = static_cast<int8_t>(omegaRef.rightSpeed);
+  sensorMsg.left_speed = omegaRef.leftSpeed;
+  sensorMsg.right_speed = omegaRef.rightSpeed;
   sensorMsg.ultrasonic_front_0 = ultrasonicData.us_front_0;
   sensorMsg.ultrasonic_front_1 = ultrasonicData.us_front_1;
   sensorMsg.ultrasonic_back = ultrasonicData.us_back;
